web-checker-requirements.txt
# SPIRITUAL WEB CHECKER REQUIREMENTS
# Ladang Berkah Digital - ZeroLight Orbit System
# Dependencies for Web Infrastructure Checker

# Core HTTP and async libraries
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# DNS and network utilities
dnspython>=2.3.0
python-whois>=0.8.0
requests>=2.28.0

# SSL and security
certifi>=2022.12.7
cryptography>=3.4.8

# Data handling and utilities
dataclasses-json>=0.5.7
python-dateutil>=2.8.2

# Optional: Enhanced features
psutil>=5.9.0          # System monitoring
colorama>=0.4.6        # Colored terminal output
rich>=13.0.0           # Rich terminal formatting
click>=8.1.0           # CLI interface

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0

# Note: Install with: pip install -r web-checker-requirements.txt